+++ cli/bench/config.go
@@ -12,6 +12,12 @@ import (
 var (
 	benchSpaceName             = "__benchmark_space__"
 	benchSpacePrimaryIndexName = "__bench_primary_key__"
+	PreFillingCount            = 1000000
+	getRandomTupleCommand      = fmt.Sprintf(
+		"box.space.%s.index.%s:random",
+		benchSpaceName,
+		benchSpacePrimaryIndexName,
+	)
 )
 
 // printConfig output formatted config parameters.
@@ -25,6 +31,10 @@ func printConfig(ctx context.BenchCtx, tarantoolConnection *tarantool.Connection
 	fmt.Printf("\tduration: %d seconds\n", ctx.Duration)
 	fmt.Printf("\tkey size: %d bytes\n", ctx.KeySize)
 	fmt.Printf("\tdata size: %d bytes\n", ctx.DataSize)
+	fmt.Printf("\tinsert: %d percentages\n", ctx.InsertCount)
+	fmt.Printf("\tselect: %d percentages\n", ctx.SelectCount)
+	fmt.Printf("\tupdate: %d percentages\n\n", ctx.UpdateCount)
+
 	fmt.Printf("Data schema\n")
 	w := tabwriter.NewWriter(os.Stdout, 1, 1, 1, ' ', 0)
 	fmt.Fprintf(w, "|\tkey\t|\tvalue\n")